requirements.txt
flask
numpy
opencv-python
pillow
torch
torchvision
segmentation-models-pytorch